Renamed matrx_core, reduced utils to public methods, new aliasing
