+++ matrx/__init__.py
@@ -1,4 +1,17 @@
-from . import *
+
+# Aliasing imports
+from matrx.actions import *
+from matrx.agents.capabilities.capability import *
+from matrx.logger import *
+from matrx.messages.message import Message
+from matrx.objects.agent_body import AgentBody
+from matrx.objects.env_object import EnvObject
+from matrx.objects.standard_objects import *
+from matrx.sim_goals import *
+from matrx.grid_world import *
+from matrx.utils import *
+from matrx.world_builder import WorldBuilder
+
 
 __docformat__ = "restructuredtext"
