+++ matrx/actions/__init__.py
@@ -1,5 +1,2 @@
-from .action import *
-from .door_actions import *
-from .move_actions import *
-from .object_actions import *
+
