+++ matrx/agents/patrolling_agent.py
@@ -1,6 +1,6 @@
-from matrx_core.agents.agent_brain import AgentBrain
-from matrx_core.agents.navigator import Navigator
-from matrx_core.agents.state_tracker import StateTracker
+from matrx.agents.agent_brain import AgentBrain
+from matrx.agents.navigator import Navigator
+from matrx.agents.state_tracker import StateTracker
 
 
 class PatrollingAgentBrain(AgentBrain):
@@ -26,7 +26,7 @@ class PatrollingAgentBrain(AgentBrain):
         return state
 
     def decide_on_action(self, state):
-        from matrx_core.agents.message import Message
+        from matrx.messages.message import Message
 
         # Send a message to a random agent
         agents = []
