+++ matrx/agents/state_tracker.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from matrx_core.utils import get_distance
+from matrx.utils import get_distance
 
 
 class StateTracker:
