+++ matrx/messages/message.py
@@ -1,7 +1,5 @@
 import json
 
-from matrx_core.utils import gen_random_string
-
 
 class Message:
     """
@@ -22,9 +20,24 @@ class Message:
         self.content = content  # content can be anything; a string, a dictionary, or even a custom object
         self.from_id = from_id  # the agent id who creates this message
         self.to_id = to_id  # the agent id who is the sender, when None it means all agents, including the sender
-        self.message_id = gen_random_string(30) # randomly generated ID of the message
+        self.message_id = __gen_random_string(30) # randomly generated ID of the message
 
     def toJSON(self):
         """ Make this class JSON serializable, such that it can be sent as JSON via the API """
         return json.dumps(self, default=lambda o: o.__dict__,
                           sort_keys=True, indent=4)
+
+
+def __gen_random_string(length):
+    """ Generates a random hexidecimal string of length 'length'.
+
+    Parameters
+    ----------
+    length
+        Length of the hexidecimal string to return.
+
+    Returns
+    -------
+        A random hexidecimal string of length 'length'.
+    """
+    return '%030x' % random.randrange(16**30)
