+++ matrx/messages/message_manager.py
@@ -1,6 +1,7 @@
-from matrx_core.agents.agent_brain import Message
+from matrx.agents.agent_brain import Message
 
-class MessageManager():
+
+class MessageManager:
     """ A manager inside the GirdWorld that tracks the received and send messages between agents and their teams.
 
         This gives provides several advantages;
