--- matrx/objects/__init__.py
@@ -1,4 +1,0 @@
-from .env_object import *
-from .standard_objects import *
-from .agent_body import *
-
