--- matrx/scenarios/defaults.json
@@ -1,40 +1,0 @@
-{
-  "simulation_defaults" :
-  {
-    "random_seed" : 1,
-    "size" : [25, 25],
-    "step_duration" : 0.1,
-    "time_focus": "step",
-    "run_visualisation_server" : false,
-    "run_matrx_api": false,
-    "simulation_goal": {
-      "goal_class": "LimitedTimeGoal",
-      "arguments":
-      {
-        "max_nr_ticks": -1
-      }
-    }
-  },
-  "AgentBody":
-  {
-    "is_traversable": false,
-    "is_movable": false,
-    "agent_speed_in_ticks": 1,
-    "visualize_size": 1.0,
-    "visualize_colour": "#92f441",
-    "visualize_opacity": 1.0,
-    "visualize_shape": 1,
-    "visualize_depth": 100,
-    "possible_actions": "*"
-  },
-  "EnvObject":
-  {
-    "is_traversable": false,
-    "is_movable": true,
-    "visualize_size": 1.0,
-    "visualize_shape": 0,
-    "visualize_colour": "#4286f4",
-    "visualize_opacity": 1.0,
-    "visualize_depth": 80
-  }
-}
